Implementing new features

+++ studysumu/base/views.py
@@ -1,11 +1,62 @@
 from django.shortcuts import render,redirect
-from .models import Room,Topic
+from .models import Room,Topic,Message
 from .form import Roomform
 from django.db.models import Q
+from django.http import HttpResponse
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.contrib.auth import authenticate,login,logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 # rooms = [
 #     {'id':1,'name':'Lets learn Python!'},
 #     {'id':2,'name':'Design with me'},
 #     {'id':3,'name':'Frontend Developers'},  ]
+
+def loginPage(request):
+
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username = username)
+        except:
+            messages.error(request,'User does not exists')
+        user = authenticate(request,username = username,password = password)
+        
+        if user is not None:
+            login(request, user)
+            return redirect ('home')
+        else:
+            messages.error(request, 'username or password does not exist')
+
+    context = {'page':page}
+    return render(request,'base/login_register.html', context)
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'An error occured in the registration')
+    return render(request,'base/login_register.html',{'form':form})
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q')!= None else ''
@@ -15,18 +66,36 @@ def home(request):
                                 Q(description__icontains=q) )
     topics = Topic.objects.all()
     room_count = rooms.count()
-    context = {'rooms':rooms,'topics':topics ,'room_count':room_count}
-    
-
-    
+    room_messages = Message.objects.all().filter(Q(room__topic__name__icontains=q))
+    context = {'rooms':rooms,'topics':topics ,'room_count':room_count,'room_messages':room_messages}
     return render(request,'base/home.html',context)
 
 def room(request,pk):
     room = Room.objects.get(id=pk)
+    room_messages = room.message_set.all()
+    participants = room.participants.all()
+    context = {'room':room,'room_messages':room_messages,'participants':participants}
     
-    context = {'room':room}
+    if request.method == 'POST':
+        message = Message.objects.create(
+        user = request.user,
+        room = room,
+        body = request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect('room',pk = room.id)
 
     return render(request,'base/room.html',context)
+
+def userProfile(request,pk):
+    user = User.objects.get(id=pk)
+    rooms = user.room_set.all()
+    room_messages = user.message_set.all()
+    topics = Topic.objects.all()
+    context = {'user':user,'rooms':rooms,'room_messages':room_messages,'topics':topics}
+    return render(request,'base/profile.html',context,)
+
+@login_required(login_url='login')
 def createRoom(request):
     if request.method == 'POST':
         print(request.POST)
@@ -39,11 +108,15 @@ def createRoom(request):
 
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     
     room = Room.objects.get(id=pk)
     form = Roomform(instance=room)
     context = {'form':form}
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!!')
 
     if request.method =='POST':
         print(request.POST)
@@ -54,10 +127,29 @@ def updateRoom(request,pk):
 
 
     return render(request,'base/room_form.html',context)
-
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!!')
+
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request,'base/delete.html',{'obj':room})
+
+@login_required(login_url='login')
+def deleteMessage(request,pk):
+    message = Message.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse('You are not allowed here!!!')
+
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    return render(request,'base/delete.html',{'obj':message})
+
